Check HTTP response in CW CountCompanies (#5)
CountCompanies() was the culprit affecting GetCompanies(), Sync and SyncController when API was not available.  This change checks the status of the HTTP call and raises an exception if it is not successful.  We may to enhance this further since the exception will get reported as an error and we may want to suppress those and just increment the metrics and go back to a waiting state.

Closes [ch565]